+++ src/route_protectors/non-auth-wrapper.tsx
@@ -1,13 +1,16 @@
 import { Outlet } from "react-router-dom";
 import NavBar from "../components/menus/publicMenu";
 import Footer from "../components/footer/Footer";
+import ScrollToTop from "../components/scrollToTop";
 
 export default function NonAuthWrapper() {
   return (
-    <main className="w-full">
-      <NavBar />
-      <Outlet />
-      <Footer />
-    </main>
+    <ScrollToTop>
+      <main className="w-full">
+        <NavBar />
+        <Outlet />
+        <Footer />
+      </main>
+    </ScrollToTop>
   );
 }