+++ src/store/index.ts
@@ -2,12 +2,20 @@ import { configureStore } from "@reduxjs/toolkit";
 import snackbar from "./app_functions/snackbar";
 import navMenuFunctions from "./app_functions/navMenuFunctions";
 
+//user
+import userAuthentication from "./user/auth";
+import myVerificationHistory from "./user/allVerificationHistory";
+
 export const projectStore = () => {
   return configureStore({
     reducer: {
       //app functionality
       snackbar: snackbar,
       menuFunctions: navMenuFunctions,
+
+      //user
+      userAuthentication: userAuthentication,
+      myVerificationHistory: myVerificationHistory,
     },
     middleware: (getDefaultMiddleware) =>
       getDefaultMiddleware({