+++ src/types/api/verificationHistory.ts
@@ -1,0 +1,11 @@
+export type verificationHistory = {
+  id: string;
+  matric_number: string;
+  timestamp: string;
+}[];
+
+export type verificationHistoryItem = {
+  id: string;
+  matric_number: string;
+  timestamp: string;
+};